PyTorch-document-AI-Model-Single-GPU-using-Vertex-Training-with-Custom-Containe-main/PyTorch document AI Model Single GPU using Vertex Training with Custom Container/requirements.txt
pip==23.1.2
accelerate==0.21.0
git+https://github.com/huggingface/transformers.git
datasets==2.16.1
pandas
typing-extensions==4.3.0
shapely
scikit-image
imgaug
pyclipper
lmdb
tqdm
numpy
visualdl
rapidfuzz
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
cython
lxml
premailer
openpyxl
attrdict
Polygon3
lanms-neo==1.0.2
PyMuPDF<1.21.0
PyYAML==6.0
fpdf==1.7.2
pdfkit==1.0.0
wkhtmltopdf==0.2
extract-msg==0.39.0
msgpack==1.0.4
pytesseract==0.3.9
seqeval==1.2.2
pdf2image==1.16.0
lxml==4.9.1
opencv-python-headless==4.6.0.66
google-cloud-vision==3.4.4
torch
torchvision
tabulate
bs4
imgkit
google-cloud-storage